tx -> EIP-2930: added generic sign(), getSenderAddress(), getSenderPublicKey(), verifySignature() tests
